+++ src/components/model/posts/Thumbnail/Thumbnail.module.css
