+++ src/components/model/posts/Thumbnail/Thumbnail.tsx
