+++ src/components/model/posts/Thumbnail/index.ts
