+++ src/components/model/posts/ThumbnailDummy/ThumbnailDummy.module.css
