+++ src/components/model/posts/ThumbnailDummy/ThumbnailDummy.tsx
