+++ src/components/model/posts/ThumbnailDummy/index.ts
