+++ src/pages/tags/[slug].tsx
@@ -3,7 +3,7 @@ import { Layout } from "components/Layout/Layout"
 import { TagDataWithPosts } from "components/interface/tag-data"
 import { List } from "components/ui/List/List"
 import { Spacer } from "components/ui/Spacer/Spacer"
-import { PostItem } from "components/page/tags/PostItem/PostItem"
+import { PostItem } from "components/model/tags/PostItem/PostItem"
 import { BlogPostRepository } from "../../infra/blog-post-repository"
 import { getSlug } from "../../utils/getSlug"
 import { SeoHeaders } from "components/ui/SeoHeaders"
